updated custom height api; added custom component to wrap sections and moved height there
