+++ demo/src/FullPage.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Pager } from "react-bootstrap";
 
-import ReactPageScroller from "../../src";
+import ReactPageScroller, { SectionContainer } from "../../src";
 import FirstComponent from "./FirstComponent";
 import SecondComponent from "./SecondComponent";
 import ThirdComponent from "./ThirdComponent";
@@ -49,7 +49,9 @@ export default class FullPage extends React.Component {
           customPageNumber={this.state.currentPage}
         >
           <FirstComponent />
-          <SecondComponent height={50}/>
+          <SectionContainer height="50%">
+            <SecondComponent />
+          </SectionContainer>
           <ThirdComponent />
           <FourthComponent />
           <FifthComponent />
